% Generated by roxygen2 (4.0.2): do not edit by hand
\name{birthdate2pin}
\alias{birthdate2pin}
\title{Given a birtdate, attach the last four digits to make a non-personal ("fake") pin}
\usage{
birthdate2pin(pos18, male_prob = 0.5)
}
\arguments{
\item{male_prob}{proportion of males in the outcome}

\item{vector}{with birtdates}
}
\value{
pin vector
}
\description{
Given a birtdate, attach the last four digits to make a non-personal ("fake") pin
}

