% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_s3_methods.R
\name{create_s3_method}
\alias{create_s3_method}
\title{Create S3 methods}
\usage{
create_s3_method(generic = NULL, object = NULL)
}
\arguments{
\item{generic, object}{arguments passed to \code{\link{NextMethod}}}
}
\value{
Function with arguments "x", "i", and "...". 
The "i" argument can be thought of as "index" if the method defined is 
for example the "["-function.
}
\description{
Helper function to create S3 method that preserves (most)
}
\keyword{internal}

