% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{fake_swedes}
\alias{fake_swedes}
\title{Fake Swede personal identity numbers and names}
\format{A data frame with 56 rows and 2 variables:
\describe{
  \item{pin}{Personal identification number, as character}
  \item{name}{Fictional Swedish names}
}}
\description{
Data set with fake personal identity numbers and names to use as example.
}

