% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/format.pin.R
\name{format}
\alias{format}
\title{Encode in a Common Format}
\usage{
format(x, ...)
}
\arguments{
\item{x}{any R object (conceptually); typically numeric. If \code{x} is of class
\code{pin} method \code{\link{format.pin}} is used, otherwise
\code{\link[base]{format}}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
This is essentialy just a wrapper to \code{\link[base]{format}}.
}
\details{
Objects of class \code{pin} inherits from class \code{AsIs}, which has its own
\code{format} method.
This function is just a wrapper to avoid calling \code{\link{format.AsIs}}.
}

