% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.oin}
\alias{is.oin}
\title{Test if a character vector contains correct \code{oin}}
\usage{
is.oin(oin)
}
\arguments{
\item{oin}{Character vector to be tested if it is an oin of the right format.}
}
\value{
Logical vector indicating if the elements can be an organisation identity number.
}
\description{
Test which elements in a text vector that contains organization identity
number.
}

