% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is.pin}
\alias{is.pin}
\title{Test if a character vector contains correct  \code{pin}}
\usage{
is.pin(pin)
}
\arguments{
\item{pin}{Character vector with swedish personal identity numbers with standard ABS format \code{"YYYYMMDDNNNC"}. See \link{pin_format}.}
}
\value{
Logical vector indicating if the elements can are of format personal identity number.
}
\description{
Test which elements of a character vector that contains correct personal
identity numbers (regarding format).
To test the pin regarding the control number use \link{pin_ctrl}.
}

