% Generated by roxygen2 (4.0.2): do not edit by hand
\name{luhn_algo}
\alias{luhn_algo}
\title{luhn_algo}
\usage{
luhn_algo(id, multiplier = c(0, 0, 2, 1, 2, 1, 2, 1, 2, 1, 2, 0))
}
\arguments{
\item{id}{Element with swedish personal identity number.}

\item{multiplier}{What should each element in id be multiplied with}
}
\value{
Logical element indicating if a pin is correct (TRUE) or not (FALSE)
}
\description{
Calculates the control number.
}
\references{
https://www.skatteverket.se/download/18.8dcbbe4142d38302d74be9/1387372677724/717B06.pdf
https://www.skatteverket.se/download/18.1e6d5f87115319ffba380001857/1285595720207/70408.pdf
}

