% Generated by roxygen2 (4.0.1): do not edit by hand
\name{oin_group}
\alias{oin_group}
\title{Calculate organization group from \code{oin}}
\usage{
oin_group(oin)
}
\arguments{
\item{oin}{Vector with swedish organization identity numbers (oin) in \code{NNNNNN-NNNN} format.}
}
\value{
Factor with organization categories.
}
\description{
Calculates the organization group from the organization number.
}
\examples{
ex_oin <- c("556000-4615", "232100-0156", "802002-4280")
oin_group(ex_oin)
}
\references{
\href{http://www.skatteverket.se/download/18.70ac421612e2a997f85800040284/1302507382017/70909.pdf}{SKV 709}
}

