% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oin_internal.R
\name{oin_group_element}
\alias{oin_group_element}
\title{oin_group_element}
\usage{
oin_group_element(one_oin)
}
\arguments{
\item{one_oin}{Character elemen with oin.}
}
\value{
Character categegory of organisational group.
}
\description{
Calculates the organization group from the organization number for one oin.
}
\references{
http://www.skatteverket.se/download/18.70ac421612e2a997f85800040284/1302507382017/70909.pdf
}

