% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pin_age}
\alias{pin_age}
\title{pin_age}
\usage{
pin_age(pin, date = Sys.Date())
}
\arguments{
\item{pin}{Character vector with personal number with format 'YYYYMMDDNNNC'. See \link{pin_format}.}

\item{date}{Date at which age is calculated.}
}
\value{
Age as an integer vector.
}
\description{
Calculate the age in full years for a given date.
}

