% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pin_birthplace_internal}
\alias{pin_birthplace_internal}
\title{pin_birthplace_internal}
\usage{
pin_birthplace_internal(pin, birth_vector)
}
\arguments{
\item{pin}{Character element with pin at standard format 'YYYYMMDDNNNC'. See \link{pin_format}.}

\item{birth_vector}{Vector mapping birth number to birthplace. See \link{pin_birthplace}.}
}
\value{
Character element containing birthplace
}
\description{
Internal computation of birthplace (one for each pin)
}

