% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin_internal.R
\name{pin_century}
\alias{pin_century}
\title{pin_century}
\usage{
pin_century(pin_short)
}
\arguments{
\item{pin_short}{Vector with swedish personal identity numbers on standard format.}
}
\value{
Century vector in numeric format
}
\description{
Assuming that the person is less than 100 years old, calculate the century of birth.
}

