% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pin_coordn}
\alias{pin_coordn}
\title{Check if \code{pin} is a coordination number}
\usage{
pin_coordn(pin)
}
\arguments{
\item{pin}{A vector of class \code{pin}. See \link{as.pin}.}
}
\value{
Logical vector indicating if the pin is a coordination number (\code{TRUE}) or pin (\code{FALSE}).
}
\description{
Calculate if the personal identity number is a coordination number.
}
\examples{
# Examples taken from SKV 704 (see references)
ex_pin <- c("196408233234", "196408833224")
pin_coordn(ex_pin)
}
\references{
\href{https://www.skatteverket.se/download/18.8dcbbe4142d38302d74be9/1387372677724/717B06.pdf}{Population registration in Sweden}
\href{https://www.skatteverket.se/download/18.1e6d5f87115319ffba380001857/1285595720207/70408.pdf}{SKV 704}
\href{http://www.riksdagen.se/sv/Dokument-Lagar/Utredningar/Statens-offentliga-utredningar/Personnummer-och-samordningsnu_GWB360/}{SOU 2008:60 : Personnummer och samordningsnummer}
}

