% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pin_coordn_correct}
\alias{pin_coordn_correct}
\title{pin_coordn_correct}
\usage{
pin_coordn_correct(pin)
}
\arguments{
\item{pin}{Vector of pins at format atandard fromat 'YYYYMMDDNNNC'. See \link{as.pin}.}
}
\value{
Character vector with pin, corrected for coordination numbers.
}
\description{
Remove the change of day in coordination numbers (to enable age calculation).
}

