% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin.R
\name{pin_date}
\alias{pin_date}
\title{Calculate the date of birth from a \code{pin}}
\usage{
pin_date(pin)
}
\arguments{
\item{pin}{A vector of class \code{pin}. See \link{as.pin}.}
}
\value{
Date of birth as a vector in date format.
}
\description{
Calculates the date of birth in date format.
}
\examples{
# Examples taken from SKV 704 (see references)
ex_pin <- c("196408233234", "186408833224")
pin_date(ex_pin)

}
