% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pin_sex}
\alias{pin_sex}
\title{pin_sex}
\usage{
pin_sex(pin)
}
\arguments{
\item{pin}{Character vector with personal number with format 'YYYYMMDDNNNC'.}
}
\value{
Factor with label 'Male' and 'Female'.
}
\description{
Calculates the sex of from the personal identification number att the format given by \code{pin_format}.
}
\references{
https://www.skatteverket.se/download/18.8dcbbe4142d38302d74be9/1387372677724/717B06.pdf
https://www.skatteverket.se/download/18.1e6d5f87115319ffba380001857/1285595720207/70408.pdf
}

