% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin.R
\name{pin_to_date-deprecated}
\alias{pin_to_date-deprecated}
\title{Calculate the date of birth from a \code{pin}}
\usage{
pin_to_date(pin)
}
\arguments{
\item{pin}{Swedish ID number}
}
\value{
Date of birth as a vector in date format.
}
\description{
Calculates the date of birth in date format.
}
\seealso{
\code{\link{sweidnumbr-deprecated}}
}
\keyword{internal}
