% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rin.R
\name{roin}
\alias{roin}
\title{Generate a vector of random \code{oin}}
\usage{
roin(n)
}
\arguments{
\item{n}{number of observations. If \code{length(n) > 1}, the length is taken to be the number required.}
}
\value{
a vector of generated \code{oin}s.
}
\description{
A function that generates random \code{oin}s (see \code{\link{as.pin}}).
The generated \code{oin} is uniformely distributed over all possible \code{oin}s.
}
\examples{
x <- roin(3)
oin_ctrl(x)
oin_group(x)

}

