% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rpin_unique}
\alias{rpin_unique}
\title{Resample duplicated subset of pin vector to make all elements unique}
\usage{
rpin_unique(pin, recursive_fun, ...)
}
\arguments{
\item{pin}{a pin vector with possibly non unique elements}

\item{recursive_fun}{function to regenerate duplicated pins}

\item{...}{arguments pased to \code{recursive_fun}}
}
\description{
Resample duplicated subset of pin vector to make all elements unique
}

