% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sweidnumbr.R
\docType{package}
\name{sweidnumbr}
\alias{sweidnumbr}
\alias{sweidnumbr-package}
\title{sweidnumbr}
\description{
Handling of swedish identity numbers.
For a quick tutorial see \code{vignette("sweidnumbr")}.
For more information see \url{https://github.com/MansMeg/sweidnumbr}.
}

