% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fake_swedes.R
\docType{data}
\name{fake_swedes}
\alias{fake_swedes}
\title{Example data}
\format{A data frame with 56 rows and 2 variables:
\describe{
  \item{pin}{identification number, as character}
  \item{name}{Fictional Swedish names}
}}
\description{
Data set with fake data to use for examples.
}

