% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oin.R
\name{is.oin}
\alias{is.oin}
\title{Test if a character vector contains correct \code{oin}}
\usage{
is.oin(oin)
}
\arguments{
\item{oin}{Character vector to be tested if it is an \code{oin} of the right format.}
}
\value{
Logical vector indicating if the elements can be an organization identity number.
}
\description{
Test which elements in a text vector that contains organization identity 
number.
}
\examples{
ex_oin <- roin(3)
is.oin(ex_oin)

ex_oin_char <- as.character(ex_oin)
is.oin(ex_oin_char)

}
