% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.oin}
\alias{is.oin}
\title{Test if a character vector contains correct \code{oin}}
\usage{
is.oin(oin)
}
\arguments{
\item{oin}{Character vector to be tested if it is an \code{oin} of the right format.}
}
\value{
Logical vector indicating if the elements can be an organization identity number.
}
\description{
Test which elements in a text vector that contains organization identity
number.
}
\examples{
ex_oin <- c("556000-4615", "232100-0156", "802002-4280", "8020024280", "AA2002-4280")
is.oin(ex_oin)
}

