% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.pin}
\alias{is.pin}
\title{is.pin}
\usage{
is.pin(pin)
}
\arguments{
\item{pin}{R object to be tested if it is a pin of the right format.}
}
\value{
Logical value indicating if the object is of class "pin"
}
\description{
Test if a vector is of class "pin".
}

