% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pin.R
\name{is.pin}
\alias{is.pin}
\title{Test if a vector is of class \code{pin}}
\usage{
is.pin(pin)
}
\arguments{
\item{pin}{A character vector to test if it is in \code{pin} format. See \link{as.pin}.}
}
\value{
Logical vector indicating if the elements can are of format personal identity number.
}
\description{
Test if a vector is of class \code{pin}
}
\examples{
ex_pin <- c("196408233234", "AA6408323234")
is.pin(ex_pin)
}

