% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pin_age}
\alias{pin_age}
\title{Calculate age of \code{pin} for a given date}
\usage{
pin_age(pin, date = Sys.Date())
}
\arguments{
\item{date}{Date at which age is calculated.}

\item{pin}{Character vector with swedish personal identity numbers with standard ABS format \code{"YYYYMMDDNNNC"}. See \link{pin_format}.}
}
\value{
Age as an integer vector.
}
\description{
Calculate the age in full years for a given date.
}

