% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pin_convert}
\alias{pin_convert}
\title{pin_convert}
\usage{
pin_convert(pin, format)
}
\arguments{
\item{pin}{A character element of length one.}

\item{format}{Which format should be converted. See \link{as.pin}.}
}
\value{
Character element on standard format.
}
\description{
Converts one pin to standard format
}

