% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pin_coordn}
\alias{pin_coordn}
\title{pin_coordn}
\usage{
pin_coordn(pin)
}
\arguments{
\item{pin}{Character vector with personal number with format 'YYYYMMDDNNNC'. See \link{pin_format}.}
}
\value{
Logical vector indicating if the pin is a coordination number (TRUE) or pin (FALSE).
}
\description{
Calculates if the personal identity number is a coordination number.
}
\references{
https://www.skatteverket.se/download/18.8dcbbe4142d38302d74be9/1387372677724/717B06.pdf
https://www.skatteverket.se/download/18.1e6d5f87115319ffba380001857/1285595720207/70408.pdf
http://www.skatteverket.se/download/18.3dfca4f410f4fc63c86800016383/70702svartvit.pdf
}

