% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pin_ctrl}
\alias{pin_ctrl}
\title{pin_ctrl}
\usage{
pin_ctrl(pin)
}
\arguments{
\item{pin}{Vector with swedish personal identity numbers (of class "pin").}
}
\value{
Logical vector indicating if a pin is correct (TRUE) or not (FALSE)
}
\description{
Calculates the control number and compare it with the control number in the personal identity number (pin).
}
\references{
https://www.skatteverket.se/download/18.8dcbbe4142d38302d74be9/1387372677724/717B06.pdf
https://www.skatteverket.se/download/18.1e6d5f87115319ffba380001857/1285595720207/70408.pdf
}

