% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pin_format}
\alias{pin_format}
\title{pin_format}
\usage{
pin_format(pin)
}
\arguments{
\item{pin}{Vector with swedish personal identity numbers in character or numeric format. See details.}
}
\value{
Character vector with swedish personal identity numbers with standard format "YYYYMMDDNNNC".
}
\description{
Converts a number of different pin formats to standard (ABS) pin format 'YYYYMMDDNNNC'.
}
\details{
The function converts different formats of swedish personal identity numbers to
the standard format. The formats that are converted are:
- numeric: YYYYMMDDNNNC
- numeric: YYMMDDNNNC
- character: "YYYYMMDDNNNC"
- character: "YYMMDD-NNNC"
- character: "YYYYMMDD-NNNC"
- character: "YYMMDDNNNC" (assuming < 100 years of age)
}
\references{
https://www.skatteverket.se/download/18.8dcbbe4142d38302d74be9/1387372677724/717B06.pdf
https://www.skatteverket.se/download/18.1e6d5f87115319ffba380001857/1285595720207/70408.pdf
}

